fixed copy weigths on windows

+++ build_model_nolearn.py
@@ -164,13 +164,14 @@ def cascade_model(options):
             print "> CNN: Loading pretrained weights from the", options['pretrained_model'], "configuration"
             pretrained_model = os.path.join(options['weight_paths'], options['pretrained_model'],'nets')
             model = os.path.join(options['weight_paths'], options['experiment'])
-            try:
-                shutil.copy(pretrained_model, model)
-            except:
-                shutil.copytree(pretrained_model, model)
-
             net1_w_def = os.path.join(model, 'nets', 'model_1.pkl')
             net2_w_def = os.path.join(model, 'nets', 'model_2.pkl')
+            
+            if not os.path.exists(model):
+                shutil.copy(pretrained_model, model)
+            else:
+                shutil.copyfile(pretrained_model, 'model_1.pkl', net1_w_def)
+                shutil.copyfile(pretrained_model, 'model_2.pkl', net2_w_def)
 
         try:
             net1.verbose = 0
